chore(ui): case-safe rename badge.tsx -> Badge.tsx and update imports
