+++ app/work/_components/WorkTimeline.client.tsx
@@ -4,7 +4,7 @@ import { motion, useScroll, useTransform } from "framer-motion";
 import { Check } from "lucide-react";
 import { useRef } from "react";
 
-import { Badge } from "@/components/ui/badge";
+import { Badge } from "@/components/ui/Badge";
 import usePrefersReducedMotion from "@/hooks/usePrefersReducedMotion";
 
 type Experience = {
